@@ -6,7 +6,7 @@ import { ServerInfo } from 'src/app/shared/models/emby/server-info';
 
 import { Component, HostListener, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { FormControl } from '@angular/forms';
-import { MatDialog } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
 
 import { Options } from '../../../shared/components/charts/options/options';
 import { NoTypeFoundDialog } from '../../../shared/dialogs/no-type-found/no-type-found.component';
@@ -25,8 +25,7 @@ import { MovieService } from '../service/movie.service';
 export class MovieOverviewComponent implements OnInit, OnDestroy {
   statistics$: Observable<MovieStatistics>;
 
-  @ViewChild(NgScrollbar)
-  textAreaScrollbar: NgScrollbar;
+  @ViewChild(NgScrollbar, {static: false}) textAreaScrollbar: NgScrollbar;
 
   selectedLibrarySub: Subscription;
   dropdownBlurredSub: Subscription;
