@@ -258,7 +258,7 @@ namespace EmbyStat.Services
 
         private UserMediaView CreateUserMediaViewFromMovie(Play play, Device device)
         {
-            var movie = _movieRepository.GetMovieById(Convert.ToInt16(play.MediaId));
+            var movie = _movieRepository.GetMovieById(play.MediaId);
             if (movie == null)
             {
                 throw new BusinessException("MOVIENOTFOUND");
@@ -285,13 +285,15 @@ namespace EmbyStat.Services
 
         private UserMediaView CreateUserMediaViewFromEpisode(Play play, Device device)
         {
-            var episode = _showRepository.GetEpisodeById(Convert.ToInt32(play.MediaId));
+            //var episode = _showRepository.GetEpisodeById(play.play.MediaId);
+            //TODO: fix implementation
+            Episode episode = null;
             if (episode == null)
             {
                 throw new BusinessException("EPISODENOTFOUND");
             }
 
-            var season = _showRepository.GetSeasonById(Convert.ToInt32(play.ParentId));
+            var season = _showRepository.GetSeasonById(play.ParentId);
             var seasonNumber = season.IndexNumber;
             var name = $"{episode.ShowName} - {seasonNumber}x{episode.IndexNumber} - {episode.Name}";
 
