@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using EmbyStat.Common.Helpers;
 using EmbyStat.Common.Models.Entities;
 using EmbyStat.Services.Models.Show;
 
@@ -10,8 +11,8 @@ namespace EmbyStat.Services.Interfaces
         IEnumerable<Library> GetShowLibraries();
         Task<ShowStatistics> GetStatistics(List<string> libraryIds);
         Task<ShowStatistics> CalculateShowStatistics(List<string> libraryIds);
-        IEnumerable<ShowCollectionRow> GetCollectedRows(List<string> libraryIds);
-        IEnumerable<ShowCollectionRow> CalculateCollectedRows(List<string> libraryIds);
+        ListContainer<ShowCollectionRow> GetCollectedRows(List<string> libraryIds, int page);
+        List<ShowCollectionRow> CalculateCollectedRows(List<string> libraryIds);
         bool TypeIsPresent();
     }
 }
