@@ -4,7 +4,7 @@ namespace EmbyStat.Services.Models.Emby
 {
     public class UserMediaView
     {
-        public int Id { get; set; }
+        public string Id { get; set; }
         public string Primary { get; set; }
         public string Name { get; set; }
         public string ParentId { get; set; }
