@@ -2,7 +2,7 @@
 {
     public class SuspiciousMovie
     {
-        public int MediaId { get; set; }
+        public string MediaId { get; set; }
         public string Title { get; set; }
         public int Number { get; set; }
     }
