@@ -2,7 +2,7 @@ import { Subscription } from 'rxjs';
 
 import { Component, Inject, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 
 import { JobService } from '../../../../shared/services/job.service';
 
