@@ -1,6 +1,12 @@
 export class ListToQueryParam {
-  public static convert(name: string, list: string[]): string {
-    let params = '?';
+  public static convert(name: string, list: string[], firstParam: boolean = true): string {
+    let params = '';
+    if (firstParam) {
+      params = '?';
+    } else {
+      params = '&';
+    }
+
     list.forEach(item => params += `${name}=${item}&`);
     return params.slice(0, -1);
   }
