@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using EmbyStat.Common.Extensions;
 using EmbyStat.Common.Models.Entities;
 using FluentAssertions;
@@ -12,11 +13,11 @@ namespace Tests.Unit.Extensions
         private readonly Show _show;
         public ShowExtensionTests()
         {
-            _show = new ShowBuilder(1, "123")
-                .AddMissingEpisodes(2, 2)
+            _show = new ShowBuilder(Guid.NewGuid().ToString(), "123")
                 .AddSeason(0, 2)
                 .AddSeason(2, 3)
                 .AddSeason(3, 2)
+                .AddMissingEpisodes(2, 2)
                 .Build();
         }
 
@@ -57,11 +58,9 @@ namespace Tests.Unit.Extensions
             missingEpisodes.Should().NotContainNulls();
             missingEpisodes.Count.Should().Be(2);
 
-            missingEpisodes[0].Id.Should().Be(0);
             missingEpisodes[0].SeasonNumber.Should().Be(2);
             missingEpisodes[0].EpisodeNumber.Should().Be(0);
 
-            missingEpisodes[1].Id.Should().Be(1);
             missingEpisodes[1].SeasonNumber.Should().Be(2);
             missingEpisodes[1].EpisodeNumber.Should().Be(1);
         }
@@ -76,7 +75,7 @@ namespace Tests.Unit.Extensions
         [Fact]
         public void NeedsShowSync_Should_Return_True_If_Show_Needs_Syncing()
         {
-            var show = new ShowBuilder(1,"1").Build();
+            var show = new ShowBuilder(Guid.NewGuid().ToString(), "1").Build();
             var needsUpdate = show.NeedsShowSync();
 
             needsUpdate.Should().BeTrue();
@@ -85,7 +84,7 @@ namespace Tests.Unit.Extensions
         [Fact]
         public void NeedsShowSync_Should_Return_True_If_Show_Failed_Sync()
         {
-            var show = new ShowBuilder(1, "1")
+            var show = new ShowBuilder(Guid.NewGuid().ToString(), "1")
                 .AddTvdbSynced(true)
                 .AddFailedSync(true).Build();
             var needsUpdate = show.NeedsShowSync();
@@ -96,7 +95,7 @@ namespace Tests.Unit.Extensions
         [Fact]
         public void NeedsShowSync_Should_Return_False_If_Show_Already_Synced()
         {
-            var show = new ShowBuilder(1, "1")
+            var show = new ShowBuilder(Guid.NewGuid().ToString(), "1")
                 .AddTvdbSynced(true).Build();
             var needsUpdate = show.NeedsShowSync();
 
@@ -106,9 +105,9 @@ namespace Tests.Unit.Extensions
         [Fact]
         public void HasShowChangedEpisode_Should_Return_True_If_Show_Has_Changed_Episodes()
         {
-            var oldShow = new ShowBuilder(1, "1").Build();
-            var newShow = new ShowBuilder(1, "1")
-                .AddEpisode(new EpisodeBuilder(1, 1, "1").Build())
+            var oldShow = new ShowBuilder(Guid.NewGuid().ToString(), "1").Build();
+            var newShow = new ShowBuilder(Guid.NewGuid().ToString(), oldShow.Id)
+                .AddEpisode(new EpisodeBuilder(Guid.NewGuid().ToString(), oldShow.Id, "1").Build())
                 .Build();
 
             var changed = newShow.HasShowChangedEpisodes(oldShow);
@@ -118,17 +117,16 @@ namespace Tests.Unit.Extensions
         [Fact]
         public void HasShowChangedEpisode_Should_Return_True_If_Show_Is_New()
         {
-            var newShow = new ShowBuilder(1, "1").Build();
+            var newShow = new ShowBuilder(Guid.NewGuid().ToString(), "1").Build();
             var changed = newShow.HasShowChangedEpisodes(null);
             changed.Should().BeTrue();
         }
 
         [Fact]
-        public void HasShowChangedEpisode_Should_Return_False_If_ShowS_Are_Equal()
+        public void HasShowChangedEpisode_Should_Return_False_If_Shows_Are_Equal()
         {
-            var oldShow = new ShowBuilder(1, "1").Build();
-            var newShow = new ShowBuilder(1, "1").Build();
-            var changed = newShow.HasShowChangedEpisodes(oldShow);
+            var oldShow = new ShowBuilder(Guid.NewGuid().ToString(), "1").Build();
+            var changed = oldShow.HasShowChangedEpisodes(oldShow);
             changed.Should().BeFalse();
         }
     }
